(enh) allow multiple conditional modmaps to apply
Resolves #86.

When searching for a conditional modmap we now also check to see
if the key pressed actually exists inside that modmap... so if
multiple modmaps match the conditional they will all be "evaluated"
until one is found that matches the key pressed.

If two conditional modmaps both match the current state and both
include the pressed key then the first one (in order of definition)
wins.  It is recommended that typically your modmaps should not
overlap like this.