(fix) holding multi-mod key down, repeat event triggers both keys!
A config like:

    multipurpose_modmap(
        "default",
        {Key.BACKSLASH: [Key.BACKSLASH, Key.RIGHT_CTRL]}
    )

If one holds down BACKSLASH the first repeat event is
treated as almost a different key and will trigger that
repeat event to fire (for the actual backspae key) but
also cause the modifier key to be output as well.

When all is said and done since the repeat output was
a glitch it will remain "incorrectly" held on the output.

This adds an extra check for suspended multi-keys that
are getting repeat events and silently drops them, as
they are not needed - they key will be "woke" either
when the multi-timer expires or when the key is released.