Make shorter aliases available for all modifiers
Since there was already "A", and "C", and the code doesn't seem to be bothered by having modifiers represented by single characters as long as they aren't in the final position in the shortcut combo, this adds short and "very short" aliases for as many modifiers as possible. 

New aliases: 
Sup, S, W, O, Sh, Cd, and their L/R variants. 

Shift and Command would be ambiguous with single character aliases, so are shortened to a minimum of 2 characters so they can still be distinguished from Super/Sup/S and Control/Ctrl/C. 

The logic of these shorter aliases should be intuitive for users that have already seen or used the "RC/LC/C" and "RA/LA/A" aliases. Leaving Meta alone for now, to avoid confusion with the past "M" = "Alt" that was removed.